Import pdfBox for Android
Had to manually add pdfBox (Tom Roush) to library and compile as it was
missing from maven. Removed standard pdfBox since it was not for
Android. Created temporary override in AppSpritzer to take all given
media and send to PDF (just for now - will revert later). Added fromURI,
getPDFtext methods to PDF.java.  Created method called openPDF in
AppSpritzer to call PDF class.

This is close to working but hangs when the PDFstripper is initialized
in the PDF class.  May be issue with library. Was successfully able to
conver the URI to a usable file.